requirements.txt
pyyaml==3.11
flask==1.0.2
requests==2.18.4